+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "unnamed-rcv-thing"
+name = "votekit"
 version = "0.1.0"
 description = ""
 authors = ["Your Name <you@example.com>"]
